[Refactor] UITableViewDataSource 분리
테이블뷰의 데이터를 제공하는 역할을 반드시 VC에서 담당할 필요는 없으므로 UITableViewDataSource를 별도 객체로 만든다.

Fixes: https://camp.yagom-academy.kr/camp/65cc2b983173745319e0464c/projects/65cc31bc3173745319e0465e
Ref: https://developer.apple.com/documentation/uikit/views_and_controls/table_views
